ccache: Make ccache aware of OpenWRT provided GCC patch.
The OpenWRT GCC patch, 910-mbsd_multi.patch adds an extra option
to gcc that depends on an environment variable. Standard ccache
is unaware of this option and therefor can produce stdout from gcc
that doesn't correspond to what you would get if you actually
do a direct compilation with gcc.

This commit adds a patch to ccache to make it aware of the new
option and removes the feature from the ccache package to use the
host system's installed ccache binary.